python/pangpuriye/pyproject.toml
[tool.poetry]
name = "pangpuriye"
version = "0.1.0"
description = ""
authors = ["Pathara Norasethasopon <patharanor@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6"
tqdm = "4.54.1"
PyYAML = "^5.3.1"
opencv-python = "^4.5.1"
torchvision = "0.5.0"
detecto = "^1.2.0"
torch = "1.4"
easyocr = "1.2.1"
matplotlib = "^3.3.3"
pytesseract = "^0.3.7"
pythainlp = "^2.2.6"
python-Levenshtein = "0.12.0"
pdf2image = "1.14.0"
strsimpy = "^0.2.0"
numpy = "1.19.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
